\name{inSilicoMerging-package}
\alias{inSilicoMerging-package}
\alias{inSilicoMerging}
\docType{package}
\title{Collection of Merging Techniques for Gene Expression Data.}

\description{
  This package provides a collection of techniques to remove inter-study bias when combining gene expression data originating from different studies.
}

\seealso{
\code{\link{merge}}
\code{\link{plotMDS}}
\code{\link{plotRLE}}
\code{\link{plotGeneWiseBoxPlot}}
}
